utils/Scripts/requirements_eval.txt
ragas
langchain-openai
datasets
httpx
pandas
python-dotenv
openai
